HttpServerModule .h+.cpp split; extract JsonSink/Sha1/Base64
KPI: 16384lights | PC:301KB | tick:35/35/98/95us(FPS:28571/28571/10204/10526) | ESP32:917KB | src:50(7407) | test:25(3240) | lizard:23w

Split src/core/HttpServerModule from a single 1294-line header into a
117-line interface header + 1023-line .cpp, so edits to the HTTP body
no longer cascade-recompile every TU that includes the header. The
header forward-declares JsonSink/Scheduler/PreviewFrame; the .cpp pulls
in the real headers. mm_core becomes a STATIC library carrying
HttpServerModule.cpp and depends on ui_embed so ui_embedded.h is
generated before compile.

Three reusable inline helpers extracted out of the old header so the
.cpp's TU isn't the only place they can live:

  src/core/JsonSink.h  - JsonSink streaming JSON writer + jsonEscape()
  src/core/Sha1.h      - RFC 3174 SHA-1 (WebSocket handshake)
  src/core/Base64.h    - RFC 4648 base64 encoder (WS + Password)

All three stay header-only: small, inline, no platform deps.

ESP32 component build (esp32/main/CMakeLists.txt) updated to list the
new .cpp — without it the link failed on the HttpServerModule vtable.
main.cpp now includes PreviewFrame.h directly (was transitively pulled
through the old header).

CLAUDE.md:
- Pre-commit policy strengthened: "Never start the checklist on your
  own — always ask first." Reviewer + ESP32 build burn ~10min of tokens;
  agents must not initiate them.
- Old "MoonModules are single-file" bullet replaced with explicit rules:
  light-domain modules + MoonModule base stay header-only; core service
  modules use .h + .cpp; exceptions need a one-line reason at file top.
- New Documentation principle: module .md is end-user / API-integrator
  doc, not tech doc. Wire contracts, cross-domain wiring, prior art,
  one mention of every control name (spec-check). Do not repeat facts
  the .h already states.

docs/moonmodules/core/HttpServerModule.md trimmed under the new
principle: dropped domain-boundary essay, redundant controls list,
implementation prose. Kept REST URL table, WS frame layouts,
cross-domain wiring note, v1/v2 prior art.

docs/plan.md: lazy-migrate FilesystemModule, NetworkModule, Scheduler,
SystemModule from header-only to .h+.cpp when next touched. MoonModule,
ModuleFactory, small utilities stay header-only.

Behaviour-neutral: ESP32 image +4 bytes vs prior commit. All 131 unit
tests pass, 8/8 scenarios, 22/22 specs, platform boundary clean.
Incremental rebuild on .cpp edit ~1.25s vs full-tree ~2.1s on header
edit. ESP32 live tick not re-captured this commit (port unplugged);
PC tick is fresh, ESP32 figure is the last committed value.
