+++ esp32/main/CMakeLists.txt
@@ -2,6 +2,7 @@ idf_component_register(
     SRCS
         "main.cpp"
         "../../src/main.cpp"
+        "../../src/core/HttpServerModule.cpp"
         "../../src/platform/esp32/platform_esp32.cpp"
     INCLUDE_DIRS
         "../../src"
